PR 2.6: io collaborator migration
Move the entire IO class (~1447 LoC) from src/pyramids/dataset/ops/io.py
into src/pyramids/dataset/_collaborators.py:IO. Migrated members:

* Public methods: read_array, write_array, to_file, to_raster,
  get_block_arrangement, get_tile, map_blocks, to_xyz,
  create_overviews, recreate_overviews, get_overview, read_overview_array.
* Public property: overview_count.
* Private helpers: _convert_polygon_to_window, _lazy_read_array,
  _read_block, _tile_offsets.

Module-level helpers in io.py stay there: _read_chunk, _write_to_file_sync,
_LAZY_IMPORT_ERROR. The IO collaborator imports the io module as
``_io_module`` and references those helpers as ``_io_module._read_chunk``
/ ``_io_module._write_to_file_sync`` so the lazy-read tests that
monkeypatch ``io_module._read_chunk`` continue to work — fresh attribute
lookup at call time, not a frozen import-time reference.

Substituted self.<state> -> self._ds.<state> for: _backend, _block_size,
_file_name, _iloc, _raster, band_count, band_names, block_size, columns,
crs, gdal_dtype, geotransform, map_to_array_coordinates, no_data_value,
numpy_dtype, raster, rows.

Three call-site fixups required:

* ``getattr(self, "_file_name", "")`` (the lazy-write file_name guard)
  was outside the regex's reach; rewritten to
  ``getattr(self._ds, "_file_name", "")``.
* ``_write_to_file_sync(self, path, ...)`` passed the IO collaborator
  where the helper expects a Dataset; switched to ``self._ds``.
* ``type(self)`` constructor / classmethod-access sites use
  ``self._ds.__class__`` (same fix as PR 2.5).

src/pyramids/dataset/dataset.py:

* Drop _IOMixin from the import alias list and from Dataset's MRO.
* Add 12 facade methods plus the overview_count facade property,
  thin-forwarding to self.io.<X>.
* Add a concrete _read_block facade (the abstract method that
  AbstractDataset declares).

src/pyramids/dataset/ops/__init__.py: drop IO. ops/io.py: keep
top-level imports and the two module-level helpers; the IO class is gone.

tests/:

* test_dataset_unit.py / test_dataset.py: tests calling private
  IO helpers (_convert_polygon_to_window, _tile_offsets) now route
  through ds.io.<X>.
* test_collaborators.py: move 12 IO method rows from FORWARDING_METHODS
  to FACADE_METHODS, move overview_count from READONLY_PROPERTIES to
  FACADE_PROPERTIES.

Full main test suite passes; only three pre-existing Windows file-lock
flakes in test_dataset_collection.py remain failing (test_to_geotiff_with_path,
test_to_geotiff_with_list_of_paths, test_to_ascii — all the same WinError 5
on the temp save dir).
