+++ src/pyramids/dataset/ops/__init__.py
@@ -1,9 +1,7 @@
 """Dataset operation mixins."""
 
 from pyramids.dataset.ops.band_metadata import BandMetadata
-from pyramids.dataset.ops.io import IO
 
 __all__ = [
     "BandMetadata",
-    "IO",
 ]
